clean up style creation
