+++ src/jsonUtils/style.js
@@ -1,11 +1,23 @@
 // @flow
-import { BorderPosition } from 'sketch-constants';
-import type { SJBorderOptions, SJShadow, SJShapeGroupLayer } from 'sketchapp-json-flow-types';
-import { makeRect, makeColorFromCSS } from './models';
-import { makeHorizontalPath, makeVerticalPath, makeShapePath, makeShapeGroup } from './shapeLayers';
-import type { Color, ViewStyle } from '../types';
+import type { SJBorderOptions, SJShadow, SJFill, SJStyle } from 'sketchapp-json-flow-types';
+import { makeColorFromCSS, makeColorFill } from './models';
+import type { ViewStyle, TextStyle } from '../types';
+import hasAnyDefined from '../utils/hasAnyDefined';
 
 const DEFAULT_SHADOW_COLOR = '#000';
+
+const SHADOW_STYLES = [
+  'shadowColor',
+  'shadowOffset',
+  'shadowOpacity',
+  'shadowRadius',
+  'shadowSpread',
+  'textShadowColor',
+  'textShadowOffset',
+  'textShadowOpacity',
+  'textShadowRadius',
+  'textShadowSpread',
+];
 
 const makeDashPattern = (style: 'dashed' | 'dotted' | 'solid', width: number): Array<number> => {
   switch (style) {
@@ -31,13 +43,30 @@ export const makeBorderOptions = (
   lineJoinStyle: 0,
 });
 
-export const makeShadow = (style: ViewStyle): SJShadow => {
-  const opacity = style.shadowOpacity !== undefined ? style.shadowOpacity : 1;
-  const color = style.shadowColor || style.color || DEFAULT_SHADOW_COLOR;
-  const radius = style.shadowRadius !== undefined ? style.shadowRadius : 1;
+export const makeShadow = (style: ViewStyle | TextStyle): SJShadow => {
+  /* eslint-disable no-nested-ternary */
+  const opacity =
+    style.shadowOpacity !== undefined
+      ? style.shadowOpacity
+      : style.textShadowOpacity !== undefined
+        ? style.textShadowOpacity
+        : 1;
+  const color = style.shadowColor || style.textShadowColor || DEFAULT_SHADOW_COLOR;
+  const radius =
+    style.shadowRadius !== undefined
+      ? style.shadowRadius
+      : style.textShadowRadius !== undefined
+        ? style.textShadowRadius
+        : 1;
   const _class = style.shadowInner !== undefined ? 'innerShadow' : 'shadow';
-  const spread = style.shadowSpread !== undefined ? style.shadowSpread : 0;
-  const { width: offsetX = 0, height: offsetY = 0 } = style.shadowOffset || {};
+  const spread =
+    style.shadowSpread !== undefined
+      ? style.shadowSpread
+      : style.textShadowSpread !== undefined
+        ? style.textShadowSpread
+        : 1;
+  const { width: offsetX = 0, height: offsetY = 0 } =
+    style.shadowOffset || style.textShadowOffset || {};
 
   return {
     _class,
@@ -55,50 +84,54 @@ export const makeShadow = (style: ViewStyle): SJShadow => {
   };
 };
 
-export const makeVerticalBorder = (
-  x: number,
-  y: number,
-  length: number,
-  thickness: number,
-  color: Color,
-): SJShapeGroupLayer => {
-  const frame = makeRect(x, y, thickness, length);
-  const shapeFrame = makeRect(0, 0, thickness, length);
-  const shapePath = makeShapePath(shapeFrame, makeVerticalPath());
-  const content = makeShapeGroup(frame, [shapePath]);
-  content.style.borders = [
-    {
-      _class: 'border',
-      isEnabled: true,
-      color: makeColorFromCSS(color),
-      fillType: 0,
-      position: BorderPosition.Center,
-      thickness,
-    },
-  ];
-  return content;
-};
+export const makeStyle = (
+  style?: ViewStyle | TextStyle,
+  fills?: Array<SJFill>,
+  shadowsProp?: Array<ViewStyle>,
+): SJStyle => {
+  const json = {
+    _class: 'style',
+    endDecorationType: 0,
+    fills: [],
+    miterLimit: 10,
+    startDecorationType: 0,
+    innerShadows: [],
+    shadows: [],
+  };
 
-export const makeHorizontalBorder = (
-  x: number,
-  y: number,
-  length: number,
-  thickness: number,
-  color: Color,
-): SJShapeGroupLayer => {
-  const frame = makeRect(x, y, length, thickness);
-  const shapeFrame = makeRect(0, 0, length, thickness);
-  const shapePath = makeShapePath(shapeFrame, makeHorizontalPath());
-  const content = makeShapeGroup(frame, [shapePath]);
-  content.style.borders = [
-    {
-      _class: 'border',
-      isEnabled: true,
-      color: makeColorFromCSS(color),
-      fillType: 0,
-      position: BorderPosition.Center,
-      thickness,
-    },
-  ];
-  return content;
+  if (fills && fills.length) {
+    json.fills = json.fills.concat(fills);
+  }
+
+  if (!style) {
+    return json;
+  }
+
+  if (style.backgroundColor) {
+    const fill = makeColorFill(style.backgroundColor);
+    json.fills.unshift(fill);
+  }
+
+  if (hasAnyDefined(style, SHADOW_STYLES)) {
+    const shadow = [makeShadow(style)];
+    if (style.shadowInner) {
+      json.innerShadows = shadow;
+    } else {
+      json.shadows = shadow;
+    }
+  }
+
+  if (shadowsProp) {
+    shadowsProp.map(shadowStyle => {
+      const shadow = makeShadow(shadowStyle);
+      if (shadowStyle.shadowInner) {
+        json.innerShadows.push(shadow);
+      } else {
+        json.shadows.push(shadow);
+      }
+      return shadowStyle;
+    });
+  }
+
+  return json;
 };
