+++ src/renderers/TextRenderer.js
@@ -27,7 +27,14 @@ export default class TextRenderer extends SketchRenderer {
     }
 
     const frame = makeRect(0, 0, layout.width, layout.height);
-    const layer = makeTextLayer(frame, name, props.textNodes, props.resizingConstraint);
+    const layer = makeTextLayer(
+      frame,
+      name,
+      props.textNodes,
+      style,
+      props.resizingConstraint,
+      props.shadows,
+    );
 
     const resolvedTextStyle = TextStyles.resolve(textStyle);
     if (resolvedTextStyle) {
