+++ src/renderers/ViewRenderer.js
@@ -1,23 +1,11 @@
 // @flow
-import { BorderPosition, FillType } from 'sketch-constants';
 import type { SJShapeGroupLayer } from 'sketchapp-json-flow-types';
 import SketchRenderer from './SketchRenderer';
-import { makeRect, makeColorFill, makeColorFromCSS } from '../jsonUtils/models';
+import { makeRect } from '../jsonUtils/models';
 import { makeRectShapeLayer, makeShapeGroup } from '../jsonUtils/shapeLayers';
 import type { ViewStyle, LayoutInfo, TextStyle } from '../types';
-import {
-  makeBorderOptions,
-  makeShadow,
-  makeHorizontalBorder,
-  makeVerticalBorder,
-} from '../jsonUtils/style';
+import { createBorders } from '../jsonUtils/borders';
 import hasAnyDefined from '../utils/hasAnyDefined';
-import same from '../utils/same';
-
-const DEFAULT_BORDER_COLOR = 'transparent';
-const DEFAULT_BORDER_STYLE = 'solid';
-
-const DEFAULT_BACKGROUND_COLOR = 'transparent';
 
 const VISIBLE_STYLES = [
   'shadowColor',
@@ -45,14 +33,6 @@ const VISIBLE_STYLES = [
 
 const OVERFLOW_STYLES = ['overflow', 'overflowX', 'overflowY'];
 
-const SHADOW_STYLES = [
-  'shadowColor',
-  'shadowOffset',
-  'shadowOpacity',
-  'shadowRadius',
-  'shadowSpread',
-];
-
 export default class ViewRenderer extends SketchRenderer {
   getDefaultGroupName() {
     return 'View';
@@ -62,38 +42,20 @@ export default class ViewRenderer extends SketchRenderer {
     layout: LayoutInfo,
     style: ViewStyle,
     textStyle: TextStyle,
-    // eslint-disable-next-line no-unused-vars
     props: any,
   ): Array<SJShapeGroupLayer> {
-    const layers = [];
+    let layers = [];
     // NOTE(lmr): the group handles the position, so we just care about width/height here
     const {
-      borderTopWidth = 0,
-      borderRightWidth = 0,
-      borderBottomWidth = 0,
-      borderLeftWidth = 0,
-
       borderTopLeftRadius = 0,
       borderTopRightRadius = 0,
       borderBottomRightRadius = 0,
       borderBottomLeftRadius = 0,
-
-      borderTopColor = DEFAULT_BORDER_COLOR,
-      borderRightColor = DEFAULT_BORDER_COLOR,
-      borderBottomColor = DEFAULT_BORDER_COLOR,
-      borderLeftColor = DEFAULT_BORDER_COLOR,
-
-      borderTopStyle = DEFAULT_BORDER_STYLE,
-      borderRightStyle = DEFAULT_BORDER_STYLE,
-      borderBottomStyle = DEFAULT_BORDER_STYLE,
-      borderLeftStyle = DEFAULT_BORDER_STYLE,
     } = style;
 
     if (!hasAnyDefined(style, VISIBLE_STYLES)) {
       return layers;
     }
-
-    const backgroundColor = style.backgroundColor || DEFAULT_BACKGROUND_COLOR;
 
     const frame = makeRect(0, 0, layout.width, layout.height);
     const radii = [
@@ -111,35 +73,7 @@ export default class ViewRenderer extends SketchRenderer {
       props.resizingConstraint,
     );
 
-    const fill = makeColorFill(backgroundColor);
-    const content = makeShapeGroup(frame, [shapeLayer], [fill]);
-
-    let innerShadows = [];
-    let shadows = [];
-
-    if (hasAnyDefined(style, SHADOW_STYLES)) {
-      const shadow = [makeShadow(style)];
-      if (style.shadowInner) {
-        innerShadows = shadow;
-      } else {
-        shadows = shadow;
-      }
-    }
-
-    if (props.shadows) {
-      props.shadows.map(shadowStyle => {
-        const shadow = makeShadow(shadowStyle);
-        if (shadowStyle.shadowInner) {
-          innerShadows.push(shadow);
-        } else {
-          shadows.push(shadow);
-        }
-        return shadowStyle;
-      });
-    }
-
-    content.style.innerShadows = innerShadows;
-    content.style.shadows = shadows;
+    const content = makeShapeGroup(frame, [shapeLayer], style, props.shadows);
 
     if (hasAnyDefined(style, OVERFLOW_STYLES)) {
       if (
@@ -154,105 +88,10 @@ export default class ViewRenderer extends SketchRenderer {
       }
     }
 
-    if (
-      same(borderTopWidth, borderRightWidth, borderBottomWidth, borderLeftWidth) &&
-      same(borderTopColor, borderRightColor, borderBottomColor, borderLeftColor) &&
-      same(borderTopStyle, borderRightStyle, borderBottomStyle, borderLeftStyle)
-    ) {
-      // all sides have same border width
-      // in this case, we can do everything with just a single shape.
-      if (borderTopStyle !== undefined) {
-        const borderOptions = makeBorderOptions(borderTopStyle, borderTopWidth);
-        if (borderOptions) {
-          content.style.borderOptions = borderOptions;
-        }
-      }
+    const contents = createBorders(content, layout, style);
 
-      if (borderTopWidth > 0) {
-        content.style.borders = [
-          {
-            _class: 'border',
-            isEnabled: true,
-            color: makeColorFromCSS(borderTopColor),
-            fillType: FillType.Solid,
-            position: BorderPosition.Inside,
-            thickness: borderTopWidth,
-          },
-        ];
-      }
-      layers.push(content);
-    } else {
-      content.hasClippingMask = true;
-      layers.push(content);
+    layers = layers.concat(contents);
 
-      if (borderTopWidth > 0) {
-        const topBorder = makeHorizontalBorder(0, 0, layout.width, borderTopWidth, borderTopColor);
-        topBorder.name = 'Border (top)';
-
-        const borderOptions = makeBorderOptions(borderTopStyle, borderTopWidth);
-        if (borderOptions) {
-          topBorder.style.borderOptions = borderOptions;
-        }
-
-        layers.push(topBorder);
-      }
-
-      if (borderRightWidth > 0) {
-        const rightBorder = makeVerticalBorder(
-          layout.width - borderRightWidth,
-          0,
-          layout.height,
-          borderRightWidth,
-          borderRightColor,
-        );
-        rightBorder.name = 'Border (right)';
-
-        const borderOptions = makeBorderOptions(borderRightStyle, borderRightWidth);
-        if (borderOptions) {
-          rightBorder.style.borderOptions = borderOptions;
-        }
-
-        layers.push(rightBorder);
-      }
-
-      if (borderBottomWidth > 0) {
-        const bottomBorder = makeHorizontalBorder(
-          0,
-          layout.height - borderBottomWidth,
-          layout.width,
-          borderBottomWidth,
-          borderBottomColor,
-        );
-        bottomBorder.name = 'Border (bottom)';
-
-        const borderOptions = makeBorderOptions(borderBottomStyle, borderBottomWidth);
-        if (borderOptions) {
-          bottomBorder.style.borderOptions = borderOptions;
-        }
-
-        layers.push(bottomBorder);
-      }
-
-      if (borderLeftWidth > 0) {
-        const leftBorder = makeVerticalBorder(
-          0,
-          0,
-          layout.height,
-          borderLeftWidth,
-          borderLeftColor,
-        );
-        leftBorder.name = 'Border (left)';
-
-        const borderOptions = makeBorderOptions(borderLeftStyle, borderLeftWidth);
-        if (borderOptions) {
-          leftBorder.style.borderOptions = borderOptions;
-        }
-
-        layers.push(leftBorder);
-      }
-
-      // TODO(lmr): how do we do transform in this case?
-    }
     return layers;
   }
 }
