+++ src/types.js
@@ -9,8 +9,6 @@ export type WrappedSketchLayer = {
   sketchObject: SketchLayer,
 };
 
-export type SketchStyle = any;
-
 export type MSArray<T> = {
   [key: number]: T,
   length: number,
@@ -23,6 +21,8 @@ export type SketchPage = {
   setName: string => void,
   layers: () => Array<SketchLayer>,
 };
+
+export type SketchStyle = any;
 
 export type SketchSharedStyleContainer = {
   setObjects: (objects: Array<SketchStyle>) => void,
@@ -83,7 +83,7 @@ export type LayoutInfo = {
   direction?: 'ltr' | 'rtl',
 };
 
-export type ViewStyle = {
+export type ViewStyle = {|
   color?: Color,
   shadowColor?: Color,
   shadowInner?: boolean,
@@ -160,15 +160,17 @@ export type ViewStyle = {
   opacity?: number,
   transform?: string,
   transformOrigin?: string,
-};
+|};
 
-export type TextStyle = {
+export type TextStyle = {|
   color?: Color,
   fontFamily?: string,
   fontSize?: number,
   fontStyle?: 'normal' | 'italic',
   fontWeight?: string,
   textDecoration?: string,
+  textShadowOpacity?: number,
+  textShadowSpread?: number,
   textShadowOffset?: { width: number, height: number },
   textShadowRadius?: number,
   textShadowColor?: Color,
@@ -178,7 +180,7 @@ export type TextStyle = {
   textAlign?: 'auto' | 'left' | 'right' | 'center' | 'justify',
   paragraphSpacing?: number,
   writingDirection?: 'auto' | 'ltr' | 'rtl',
-};
+|};
 
 export type TextNode = { content: string, textStyles: TextStyle };
 export type TextNodes = Array<TextNode>;
