+++ src/utils/constants.js
@@ -18,9 +18,11 @@ export const INHERITABLE_FONT_STYLES = [
   'fontWeight',
   'textAlign',
   'textDecoration',
-  'textShadowOffset',
-  'textShadowRadius',
   'textShadowColor',
+  'textShadowOffset',
+  'textShadowOpacity',
+  'textShadowRadius',
+  'textShadowSpread',
   'textTransform',
   'letterSpacing',
   'lineHeight',
